fix: lazy load workspace export workflow
